djangobackend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.8.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
sqlparse==0.5.3
typing_extensions==4.14.1
